Move fixtures up a folder
so we can share then between unit and functional tests
