+++ tests/unit/docker_images/test_checksummer.py
@@ -4,7 +4,8 @@ import uuid
 
 import boto3
 
-from .. import UploadTestCaseUsingMockAWS, EnvironmentSetup, FIXTURE_DATA_CHECKSUMS
+from .. import UploadTestCaseUsingMockAWS, EnvironmentSetup
+from ... import FIXTURE_DATA_CHECKSUMS
 
 # pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'docker-images', 'checksummer'))  # noqa
 # sys.path.insert(0, pkg_root)  # noqa
