+++ tests/unit/lambdas/test_checksum_daemon.py
@@ -7,7 +7,7 @@ import uuid
 import boto3
 
 from .. import UploadTestCaseUsingMockAWS, EnvironmentSetup
-from .. import FIXTURE_DATA_CHECKSUMS
+from ... import FIXTURE_DATA_CHECKSUMS
 
 from upload.common.upload_area import UploadArea
 from upload.common.upload_config import UploadConfig
